partial fix reviewer comment

+++ src/kernels/MIOpenLossReduce.cpp
@@ -32,7 +32,7 @@
 #include "warp_shuffle.hpp"
 
 template <typename DTYPE>
-__device__ void LossSum(const DTYPE* __restrict__ input, DTYPE* __restrict__ output, size_t N)
+__device__ void LossSum(const DTYPE* __restrict__ input, DTYPE* __restrict__ output, uint64_t N)
 {
     auto gid = blockIdx.x * blockDim.x + threadIdx.x;
 
@@ -44,7 +44,7 @@ __device__ void LossSum(const DTYPE* __restrict__ input, DTYPE* __restrict__ out
 }
 
 extern "C" __global__ void
-ReduceSumLoss(const FLOAT* __restrict__ input, FLOAT* __restrict__ output, size_t N)
+ReduceSumLoss(const FLOAT* __restrict__ input, FLOAT* __restrict__ output, uint64_t N)
 {
     // instantiate the kernel
     LossSum<FLOAT>(input, output, N);
